% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_analysis.R
\name{vector_analysis}
\alias{vector_analysis}
\title{Vector Analysis Function}
\usage{
vector_analysis(x)
}
\arguments{
\item{x}{A numeric vector to analyze.}
}
\value{
A result of vector analysis.
}
\description{
This function performs custom vector analysis.
}
\examples{
vector_analysis(c(1, 2, 3))
}
